fourth version

+++ hotel_service/src/main/java/com/hotel/service/ClientServices.java
@@ -3,20 +3,24 @@ package com.hotel.service;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import com.hotel.api.services.IClientServices;
 import com.hotel.api.storage.IClientDAO;
 import com.hotel.entities.Client;
 import com.hotel.storages.ClientDAO;
 
 public class ClientServices implements IClientServices {
+	private static final Logger logger = LoggerFactory.getLogger(ClientServices.class);
 	IClientDAO clientDAO = new ClientDAO();
-
+	
 	@Override
 	public void addClient(Client client) {
 		if (!clientDAO.listAllClients().contains(client)) {
 			clientDAO.addClient(client);
 		} else {
-			System.out.println("This client is already exists");
+			logger.info("This client is already exists");
 		}
 	}
 
@@ -27,7 +31,7 @@ public class ClientServices implements IClientServices {
 
 	@Override
 	public List<String> listAllNameClients() {
-		return clientDAO.listAllClients().stream().map(client -> client.getClientName()).collect(Collectors.toList());
+		return clientDAO.listAllClients().stream().map(Client::getClientName).collect(Collectors.toList());
 	}
 
 	@Override
@@ -38,8 +42,12 @@ public class ClientServices implements IClientServices {
 
 	@Override
 	public void removeClient(Client client) {
-		clientDAO.removeClient(client);
-
+		if (clientDAO.listAllClients().contains(client)) {
+			clientDAO.removeClient(client);
+		} else {
+			logger.info("This client isn't exists");
+		}
+		
 	}
 
 }
